+++ CHANGELOG.md
@@ -16,6 +16,8 @@ All notable changes to the "robotcode" extension will be documented in this file
     - fixes [#20](https://github.com/d-biehl/robotcode/issues/20)
     - note: i think you should not use colons or semicolon in testcase names ;-)
 - Fix stepping/halt on breakpoint for IF/ELSE statements if the expression is evaluated as False
+- Rework of stepping and stacktrace in the debugger
+  - Only the real keyword calls are displayed in the stack trace.
 
 ##  0.5.5